Simple_Fun_#154-Zero_And_One
You have a string of length n consisting of zeroes and ones. Consider the following operation:

Choose any two adjacent positions in the string
If one of them is 0, and the other one is 1,
remove these two digits from the string.
What is the length of the smallest string that you can get after applying this operation multiple times?

+Input/Output
[input] string s
The initial string.

[output] an integer
The minimum length of the string that may remain after applying the described operations several times.